# Conflicts:
#	.gitignore
#	app/Http/Controllers/AccountController.php
#	app/Http/Controllers/QuoteApiController.php
#	app/Http/Requests/CreateClientRequest.php
#	app/Http/Requests/CreateCreditRequest.php
#	app/Http/Requests/CreateInvoiceRequest.php
#	app/Http/Requests/CreatePaymentTermRequest.php
#	app/Http/Requests/CreateVendorRequest.php
#	app/Http/Requests/UpdateInvoiceRequest.php
#	app/Http/Requests/UpdateVendorRequest.php
#	app/Http/routes.php
#	app/Models/Account.php
#	public/built.js
#	resources/lang/es_ES/texts.php
#	resources/lang/it/texts.php
#	resources/views/invoices/edit.blade.php
#	resources/views/master.blade.php

+++ app/Ninja/Repositories/ExpenseRepository.php
@@ -1,13 +1,13 @@
-<?php namespace App\Ninja\Repositories;
+<?php
 
-use DB;
-use Utils;
-use Auth;
+namespace App\Ninja\Repositories;
+
+use App\Models\Document;
 use App\Models\Expense;
 use App\Models\Vendor;
-use App\Models\Document;
-use App\Ninja\Repositories\BaseRepository;
-use Session;
+use Auth;
+use DB;
+use Utils;
 
 class ExpenseRepository extends BaseRepository
 {
@@ -36,22 +36,8 @@ class ExpenseRepository extends BaseRepository
     public function findVendor($vendorPublicId)
     {
         $vendorId = Vendor::getPrivateId($vendorPublicId);
-        $accountid = \Auth::user()->account_id;
-        $query = DB::table('expenses')
-                    ->join('accounts', 'accounts.id', '=', 'expenses.account_id')
-                    ->where('expenses.account_id', '=', $accountid)
-                    ->where('expenses.vendor_id', '=', $vendorId)
-                    ->select(
-                        'expenses.id',
-                        'expenses.expense_date',
-                        'expenses.amount',
-                        'expenses.public_notes',
-                        'expenses.public_id',
-                        'expenses.deleted_at',
-                        'expenses.should_be_invoiced',
-                        'expenses.created_at',
-                        'expenses.user_id'
-                    );
+
+        $query = $this->find()->where('expenses.vendor_id', '=', $vendorId);
 
         return $query;
     }
@@ -65,21 +51,23 @@ class ExpenseRepository extends BaseRepository
                     ->leftJoin('contacts', 'contacts.client_id', '=', 'clients.id')
                     ->leftjoin('vendors', 'vendors.id', '=', 'expenses.vendor_id')
                     ->leftJoin('invoices', 'invoices.id', '=', 'expenses.invoice_id')
+                    ->leftJoin('expense_categories', 'expenses.expense_category_id', '=', 'expense_categories.id')
                     ->where('expenses.account_id', '=', $accountid)
                     ->where('contacts.deleted_at', '=', null)
                     ->where('vendors.deleted_at', '=', null)
                     ->where('clients.deleted_at', '=', null)
-                    ->where(function ($query) {
+                    ->where(function ($query) { // handle when client isn't set
                         $query->where('contacts.is_primary', '=', true)
                               ->orWhere('contacts.is_primary', '=', null);
                     })
                     ->select(
-                        DB::raw('COALESCE(expenses.invoice_id, expenses.should_be_invoiced) expense_status_id'),
+                        DB::raw('COALESCE(expenses.invoice_id, expenses.should_be_invoiced) status'),
                         'expenses.account_id',
                         'expenses.amount',
                         'expenses.deleted_at',
                         'expenses.exchange_rate',
-                        'expenses.expense_date',
+                        'expenses.expense_date as expense_date_sql',
+                        DB::raw("CONCAT(expenses.expense_date, expenses.created_at) as expense_date"),
                         'expenses.id',
                         'expenses.is_deleted',
                         'expenses.private_notes',
@@ -91,8 +79,15 @@ class ExpenseRepository extends BaseRepository
                         'expenses.expense_currency_id',
                         'expenses.invoice_currency_id',
                         'expenses.user_id',
+                        'expenses.tax_rate1',
+                        'expenses.tax_rate2',
+                        'expenses.payment_date',
+                        'expense_categories.name as category',
+                        'expense_categories.user_id as category_user_id',
+                        'expense_categories.public_id as category_public_id',
                         'invoices.public_id as invoice_public_id',
                         'invoices.user_id as invoice_user_id',
+                        'invoices.balance',
                         'vendors.name as vendor_name',
                         'vendors.public_id as vendor_public_id',
                         'vendors.user_id as vendor_user_id',
@@ -105,17 +100,47 @@ class ExpenseRepository extends BaseRepository
                         'clients.country_id as client_country_id'
                     );
 
-        $showTrashed = \Session::get('show_trash:expense');
+        $this->applyFilters($query, ENTITY_EXPENSE);
 
-        if (!$showTrashed) {
-            $query->where('expenses.deleted_at', '=', null);
+        if ($statuses = session('entity_status_filter:' . ENTITY_EXPENSE)) {
+            $statuses = explode(',', $statuses);
+            $query->where(function ($query) use ($statuses) {
+                $query->whereNull('expenses.id');
+
+                if (in_array(EXPENSE_STATUS_LOGGED, $statuses)) {
+                    $query->orWhere('expenses.invoice_id', '=', 0)
+                          ->orWhereNull('expenses.invoice_id');
+                }
+                if (in_array(EXPENSE_STATUS_INVOICED, $statuses)) {
+                    $query->orWhere('expenses.invoice_id', '>', 0);
+                    if (! in_array(EXPENSE_STATUS_BILLED, $statuses)) {
+                        $query->where('invoices.balance', '>', 0);
+                    }
+                }
+                if (in_array(EXPENSE_STATUS_BILLED, $statuses)) {
+                    $query->orWhere('invoices.balance', '=', 0)
+                          ->where('expenses.invoice_id', '>', 0);
+                }
+                if (in_array(EXPENSE_STATUS_PAID, $statuses)) {
+                    $query->orWhereNotNull('expenses.payment_date');
+                }
+                if (in_array(EXPENSE_STATUS_UNPAID, $statuses)) {
+                    $query->orWhereNull('expenses.payment_date');
+                }
+                if (in_array(EXPENSE_STATUS_PENDING, $statuses)) {
+                    $query->orWhere('expenses.should_be_invoiced', '=', 1)
+                            ->whereNull('expenses.invoice_id');
+                }
+            });
         }
 
         if ($filter) {
             $query->where(function ($query) use ($filter) {
                 $query->where('expenses.public_notes', 'like', '%'.$filter.'%')
                       ->orWhere('clients.name', 'like', '%'.$filter.'%')
-                      ->orWhere('vendors.name', 'like', '%'.$filter.'%');
+                      ->orWhere('vendors.name', 'like', '%'.$filter.'%')
+                      ->orWhere('expense_categories.name', 'like', '%'.$filter.'%');
+                ;
             });
         }
 
@@ -130,42 +155,50 @@ class ExpenseRepository extends BaseRepository
             // do nothing
         } elseif ($publicId) {
             $expense = Expense::scope($publicId)->firstOrFail();
-            \Log::warning('Entity not set in expense repo save');
+            if (Utils::isNinjaDev()) {
+                \Log::warning('Entity not set in expense repo save');
+            }
         } else {
             $expense = Expense::createNew();
+        }
+
+        if ($expense->is_deleted) {
+            return $expense;
         }
 
         // First auto fill
         $expense->fill($input);
 
-        $expense->expense_date = Utils::toSqlDate($input['expense_date']);
-
-        if (isset($input['private_notes'])) {
-            $expense->private_notes = trim($input['private_notes']);
+        if (isset($input['expense_date'])) {
+            $expense->expense_date = Utils::toSqlDate($input['expense_date']);
         }
-        $expense->public_notes = trim($input['public_notes']);
-        $expense->should_be_invoiced = isset($input['should_be_invoiced']) || $expense->client_id ? true : false;
+        if (isset($input['payment_date'])) {
+            $expense->payment_date = Utils::toSqlDate($input['payment_date']);
+        }
 
-        if ( ! $expense->expense_currency_id) {
+        if (! $expense->expense_currency_id) {
             $expense->expense_currency_id = \Auth::user()->account->getCurrencyId();
         }
-        if ( ! $expense->invoice_currency_id) {
+        if (! $expense->invoice_currency_id) {
             $expense->invoice_currency_id = \Auth::user()->account->getCurrencyId();
         }
 
         $rate = isset($input['exchange_rate']) ? Utils::parseFloat($input['exchange_rate']) : 1;
         $expense->exchange_rate = round($rate, 4);
-        $expense->amount = round(Utils::parseFloat($input['amount']), 2);
+        if (isset($input['amount'])) {
+            $expense->amount = round(Utils::parseFloat($input['amount']), 2);
+        }
 
         $expense->save();
 
         // Documents
-        $document_ids = !empty($input['document_ids'])?array_map('intval', $input['document_ids']):array();;
-        foreach ($document_ids as $document_id){
+        $document_ids = ! empty($input['document_ids']) ? array_map('intval', $input['document_ids']) : [];
+
+        foreach ($document_ids as $document_id) {
             // check document completed upload before user submitted form
             if ($document_id) {
                 $document = Document::scope($document_id)->first();
-                if($document && Auth::user()->can('edit', $document)){
+                if ($document && Auth::user()->can('edit', $document)) {
                     $document->invoice_id = null;
                     $document->expense_id = $expense->id;
                     $document->save();
@@ -173,55 +206,16 @@ class ExpenseRepository extends BaseRepository
             }
         }
 
-        if(!empty($input['documents']) && Auth::user()->can('create', ENTITY_DOCUMENT)){
-            // Fallback upload
-            $doc_errors = array();
-            foreach($input['documents'] as $upload){
-                $result = $this->documentRepo->upload($upload);
-                if(is_string($result)){
-                    $doc_errors[] = $result;
+        // prevent loading all of the documents if we don't have to
+        if (! $expense->wasRecentlyCreated) {
+            foreach ($expense->documents as $document) {
+                if (! in_array($document->public_id, $document_ids)) {
+                    // Not checking permissions; deleting a document is just editing the invoice
+                    $document->delete();
                 }
-                else{
-                    $result->expense_id = $expense->id;
-                    $result->save();
-                    $document_ids[] = $result->public_id;
-                }
-            }
-            if(!empty($doc_errors)){
-                Session::flash('error', implode('<br>',array_map('htmlentities',$doc_errors)));
-            }
-        }
-
-        foreach ($expense->documents as $document){
-            if(!in_array($document->public_id, $document_ids)){
-                // Not checking permissions; deleting a document is just editing the invoice
-                $document->delete();
             }
         }
 
         return $expense;
-    }
-
-    public function bulk($ids, $action)
-    {
-        $expenses = Expense::withTrashed()->scope($ids)->get();
-
-        foreach ($expenses as $expense) {
-            if ($action == 'restore') {
-                $expense->restore();
-
-                $expense->is_deleted = false;
-                $expense->save();
-            } else {
-                if ($action == 'delete') {
-                    $expense->is_deleted = true;
-                    $expense->save();
-                }
-
-                $expense->delete();
-            }
-        }
-
-        return count($tasks);
     }
 }
